Add autoconf support
Introduce autoconf configure script and makefile as an alternative to
the default building method.

- It is possible now to build the McDst and converters for Darwin (MacOS)
  systems.

- configure script allows to enable/disable UrQMD and/or Pythia 8
  converters. By default, configure script, enables UrQMD and disables
  Pythia 8.

To configure the Makefile.auto.in template one could use the following
options:

for UrQMD:
--enable-urqmd     enable UrQMD converter

for Pythia8:
--enable-pythia8   enable pythia8 converter
The library searching is doing with helper script pythia8-config which
is provded by the pythia8 source code. By default this script are
searching in the $PATH environment variable, but it is possible to
override this behaviour by the following option:
--with-pythia8=<ARG>
where <ARG> is a path to the directory which contains the pythia8-config
script.

Debug:
By default the main library and converters is built without debuginfo
with 2nd level compiler optimizations. To enable it, one could use the
following options:
--enable-debug     enable -g -O0
--enable-opt-debug enable -g -O2
One option is for DWARF debuginfo without compiler optimizations and
another one is for DWARF debuginfo with compiler 2nd level
optimizations.

Brief and simple bootstrap and build instruction:

```sh
$ # bootstrapping
$ autoconf
$ ./configure
$ make -f Makefile.auto
$ make -f Makefile.auto converters
```